+++ Models/Team.cs
@@ -3,7 +3,7 @@
     public class Team
     {
         public int Id { get; set; }
-        public int ApiTeamId { get; set; }
+        public long ApiTeamId { get; set; }
         public string Name { get; set; } = string.Empty;
         public string ShortName { get; set; } = string.Empty;
         public string Country { get; set; } = string.Empty;
